// Zelnak-eshop/frontend/src/enums/Routes.ts
export enum Routes {
    Homepage = '/',
    Register = '/register',
    Login = '/login',
    Products = '/products',
    Seller = '/farmers',
    Profile = '/profile',
    AddProduct = '/add-product',
    Events = '/events',
    Categories = '/categories',
    AddCategory = '/add-category',
    Cart = '/cart',
    Orders = '/orders',
    FarmerProfile = '/farmers',
    AdminPage = '/admin',
    AddEvent = '/add-event',
    EditEvent = '/edit-event',
    EditProduct = '/edit-product',
}